requirements.txt
comet_ml==3.53.2
numpy==2.3.4
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==12.0.0
PyYAML==6.0.2
PyYAML==6.0.3
scipy==1.16.2
skimage==0.0
submitit==1.5.3
torch==2.1.2+cu118
torchdiffeq==0.2.5
torchmetrics==1.8.2
torchvision==0.16.2+cu118
tqdm==4.67.1
